WIP to allow views to be used standalone.
Right now, the library doesn't really support our use-case:
we want the views on their own so we can do a two-step enrolment
during user setup. Unfortunately the views are tied into a
delegate during instantiation, and can't be instantiated from a
storyboard/NIB because they lack -initWithFrame. This is a quick
hack to allow us to reject the -dismissViewController call after
a PIN is accepted.